Remove default setting of AgentImage from traffic-manager's environment
The agent-image should ideally be determined by asking SystemA, and the
environment variable should be used if SystemA isn't available, i.e.
when the system is air-gapped. A default setting must only be used when
SystemA is unavailable and the environment variable is unset, therefore
the default setting of the environment variable is removed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>