Fetch extended agent-image from SystemA
When no agent-image has been explicitly declared in the traffic-manager
environment, then an attempt will be made to fetch this image from
SystemA. If that attempt fails and the auto-injector will use the
default `tel2:<version>` image to resolve `create` requests and the
`PrepareIntercept` will error unless the protocol is `tcp`.

After a successful retrieval of the `agent-image`, is assumed to be able
to handle all mechanisms. The agent-image then and becomes persistent
and will be used for all protocols (including `tpc`, which is fine
because the smart-agent is now capable of doing that).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>